[network] Mockup pipelined SNN implementation
This patch adds a new class, 'network', to distribute and manage
the pipelined execution of a spiking neural network.  We begin
by considering only layer-wise parallelism and do not define a
load-balancing algorithm, among many other key features.

This is, therefore, more of a conceptual base for future work to
be built upon.